refactor(utils): restructure
