+++ src/.lintstagedrc
