+++ src/utils/networking.ts
@@ -1,3 +1,21 @@
+import { type H3Event } from 'h3'
+
+export const getDomainTldPort = (host: string) => {
+  const hostParts = host.split('.')
+
+  if (hostParts.length <= 2) return host
+
+  return `${hostParts[hostParts.length - 2]}.${hostParts[hostParts.length - 1]}`
+}
+
+export const getHost = (event: H3Event) => {
+  const host = event.node.req.headers.host
+
+  if (!host) throw new Error('Host header is not given!')
+
+  return host
+}
+
 export const getServiceHref = ({
   host,
   isSsr = true,
