+++ src/utils/types.ts
@@ -1,0 +1,3 @@
+export const isNeitherNullNorUndefined = <T>(
+  value: T | null | undefined,
+): value is T => value !== null && value !== undefined
