+++ tests/.lintstagedrc
@@ -1,0 +1,3 @@
+{
+  "*.{js,ts}": "eslint"
+}
